pyproject.toml
[tool.poetry]
name = "py-image-ascii"
version = "0.1.0"
description = ""
readme = "README.md"
license = "MIT"
authors = ["David Buckley <buckley.w.david@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^6.7"
pillow = "^5.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
mypy = "^0.620.0"
pytest-cov = "^2.5"
black = "^18.3-alpha.0"

[tool.poetry.scripts]
img2ascii = 'py_image_ascii.cli:main'
